+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Option to run a mirror of smoke.io
+- Git to `Dockerfile`
+- 28080 for webapp to firewall allows in `install.sh`
+
+### Changed
+- `changelog.md`
+
+### Removed
+- Option to add custom ports tie into docker.
 
 ## 2.2.0 - 2020-02-29
 ### Added